requirements.txt
certifi
chardet
click
decorator
Flask
gunicorn
idna
ipython-genutils
itsdangerous
Jinja2
jsonschema
jupyter-core
MarkupSafe
nbformat
numpy
pandas
pipenv
plotly
python-dateutil
pytz
requests
six
traitlets
urllib3
virtualenv
virtualenv-clone
Werkzeug
sqlalchemy
networkx
gensim
pillow
wordcloud